v1.0.2 - added x86 linux support - fixed ogg vorbis seeking bug
git-svn-id: https://www.daphne-emu.com:9443/daphnesvn/branches/v_1_0@2872 c98d566b-fa28-4d4c-8a7b-932307ef5c1a

+++ src/ldp-in/ldv1000.cpp
@@ -607,9 +607,9 @@ unsigned int g_ldv1000_last_event = 0;
 
 void ldv1000_event_callback(void *eventType)
 {
-	g_ldv1000_last_event = (unsigned int) eventType;
+	g_ldv1000_last_event = (unsigned long) eventType;	// changed to long for x64 support (thanks Joker)
 
-	switch ((unsigned int) eventType)
+	switch ((unsigned long) eventType)
 	{
 	case LDV1000_EVENT_VSYNC_END:
 #ifdef DEBUG
